added code

+++ src/main/java/com/example/AIRBNBCLONE/service/RoomServiceImpl.java
@@ -3,11 +3,14 @@ package com.example.AIRBNBCLONE.service;
 import java.util.List;
 
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.example.AIRBNBCLONE.dto.RoomDto;
 import com.example.AIRBNBCLONE.entity.Hotel;
 import com.example.AIRBNBCLONE.entity.Room;
+import com.example.AIRBNBCLONE.entity.User;
+import com.example.AIRBNBCLONE.exception.UnAuthorisedException;
 import com.example.AIRBNBCLONE.repository.HotelRepository;
 import com.example.AIRBNBCLONE.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,43 +21,61 @@ public class RoomServiceImpl implements RoomService {
     private final ModelMapper modelMapper;
     private final RoomRepository roomRepository;
     private final HotelRepository hotelRepository;
+    private final InventoryService inventoryService;
 
     @Override
     public RoomDto createRoom(RoomDto roomDto, Long hotelId) {
-       Hotel hotel=hotelRepository.findById(hotelId).orElseThrow(()->new IllegalArgumentException("Hotel not found with Id: "+hotelId));
-       Room room=modelMapper.map(roomDto, Room.class);
-       room.setHotel(hotel);
-       Room savedRoom = roomRepository.save(room);
-       return modelMapper.map(savedRoom, RoomDto.class);
+        Hotel hotel = hotelRepository.findById(hotelId)
+                .orElseThrow(() -> new IllegalArgumentException("Hotel not found with Id: " + hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("User Is Invalid to create room");
+        }
+        Room room = modelMapper.map(roomDto, Room.class);
+        room.setHotel(hotel);
+        Room savedRoom = roomRepository.save(room);
+
+        if (hotel.getActive()) {
+            inventoryService.initializeRoomForAYear(room);
+        }
+        return modelMapper.map(savedRoom, RoomDto.class);
     }
 
     @Override
     public RoomDto getRoomById(Long roomId) {
-        Room room=roomRepository.findById(roomId).orElseThrow(()->new IllegalArgumentException("Room not found with Id: "+roomId));
+        Room room = roomRepository.findById(roomId)
+                .orElseThrow(() -> new IllegalArgumentException("Room not found with Id: " + roomId));
         return modelMapper.map(room, RoomDto.class);
     }
 
     @Override
     public void deleteRoomById(Long roomId) {
-        Room room=roomRepository.findById(roomId).orElseThrow(()->new IllegalArgumentException("Room not found with Id: "+roomId));
-         roomRepository.delete(room);
+        Room room = roomRepository.findById(roomId)
+                .orElseThrow(() -> new IllegalArgumentException("Room not found with Id: " + roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this room with id: " + roomId);
+        }
+        inventoryService.deleteAllInventory(room);
+        roomRepository.delete(room);
     }
 
     @Override
     public RoomDto updateRoomById(Long roomId, RoomDto roomDto) {
-        Room existingRoom=roomRepository.findById(roomId).orElseThrow(()->new IllegalArgumentException("Room not found with Id: "+roomId));
+        Room existingRoom = roomRepository.findById(roomId)
+                .orElseThrow(() -> new IllegalArgumentException("Room not found with Id: " + roomId));
         modelMapper.map(roomDto, existingRoom);
         existingRoom.setId(roomId);
-        Room updatedRoom=roomRepository.save(existingRoom);
+        Room updatedRoom = roomRepository.save(existingRoom);
         return modelMapper.map(updatedRoom, RoomDto.class);
     }
 
     @Override
     public List<RoomDto> getAllRoomsByHotelId(Long hotelId) {
-        Hotel hotel=hotelRepository.findById(hotelId).orElseThrow(()->new IllegalArgumentException("Hotel not found with Id: "+hotelId));
-        List<Room> rooms=roomRepository.findByHotel(hotel);
-        return rooms.stream().map(room->modelMapper.map(room, RoomDto.class)).collect(java.util.stream.Collectors.toList());
+        Hotel hotel = hotelRepository.findById(hotelId)
+                .orElseThrow(() -> new IllegalArgumentException("Hotel not found with Id: " + hotelId));
+        List<Room> rooms = roomRepository.findByHotel(hotel);
+        return rooms.stream().map(room -> modelMapper.map(room, RoomDto.class))
+                .collect(java.util.stream.Collectors.toList());
     }
-    }
-    
-
+}
